+++ R/BDD.ouverture.R
@@ -63,6 +63,7 @@ BDD.ouverture <- function(
     if(system('uname -n',intern=T) == "MacBookJB"){utilisateur <- "jb"}
     if(system('uname -n',intern=T) == "MBPdeJeBaptiste"){utilisateur <- "jb"}
     if(system('uname -n',intern=T) == "MBPdeJeBaptiste.cloud.peche-jura.com"){utilisateur <- "jb"}
+    if(system('uname -n',intern=T) == "MacBook-Pro-de-Jean-Baptiste.local"){utilisateur <- "jb"}
     if(system('uname -n',intern=T) == "MBP-de-Adrien"){utilisateur <- "adrien"}
     if(system('uname -n',intern=T) == "MBP-de-Adrien.local"){utilisateur <- "adrien"}
     if(system('uname -n',intern=T) == "MBP-de-Adrien.cloud.peche-jura.com"){utilisateur <- "adrien"}