+++ R/chronique.periode.R
@@ -2,19 +2,15 @@
 #'
 #' Cette fonction permet de filtrer la période d'intérêt. La valeur de début est l'heure sélectionnée +1, la valeur de fin est strictement inférieure à celle indiquée
 #' @name chronique.periode
-#' @param Début Date de début (2015-02-01 14:00:00)
-#' @param Fin Date de fin (2015-02-04 14:00:00)
+#' @param deb Date de début (2015-02-01 14:00:00)
+#' @param fin Date de fin (2015-02-04 14:00:00)
 #' @keywords chronique
 #' @import tidyverse
 #' @export
 #' @examples
-#' chronique.periode(Data, "2015-02-04 11:00:00", )
-#' chronique.periode(Data, ,"2015-02-04 14:00:00")
+#' chronique.periode(Data, "2015-02-04 11:00:00")
+#' chronique.periode(Data, "2015-02-04 14:00:00")
 #' chronique.periode(Data, "2015-02-04 11:00:00", "2015-02-04 14:00:00")
-
-##### TODO LIST #####
-# 
-#####################
 
 chronique.periode <- function(x = Data, deb = "", fin = "")