+++ R/chronique.resultats.periode.R
@@ -3,20 +3,14 @@
 #' Cette fonction permet de calculer les périodes couvertes par les résultats de chroniques (champs PeriodeTotale)
 #' @name chronique.resultats.periode
 #' @param data Data.frame issu de la fonction chronique.resultats ou de la base de données Chroniques
-#' @param typeperiode Type de période que l'on souhaite calculer : période totale (\code{Enveloppe} : 2014-2018), ou période fractionnée (\code{Detail} : 2014-2015;2017;2019-2020, par défaut)
+#' @param typeperiode Type de période que l'on souhaite calculer : période totale (\code{Enveloppe} : 2014-2018), ou période fractionnée (\code{Detail} : 2014-2015,2017,2019-2020, par défaut)
 #' @keywords chronique
+#' @import glue
 #' @import tidyverse
 #' @export
 #' @examples
 #' Resultats %>% chronique.resultats.periode()
 #' chronique.resultats("2017","Année biologique") %>% chronique.resultats.periode()
-
-##### TODO LIST #####
-# Vérifier à quoi ressemble le cas où il n'y a qu'une année, et faire pour que cette année seule apparaisse plutôt que 2020-2020
-# Implanter cette fonction dans chronique.analyse() ou chronique.traitement() plutôt, afin que le calcul de l'enveloppe soit directement juste, notamment pour les autres FD
-# Créer une fonction plus générale pour le renommage initial des champs, à ensuite également utiliser dans chronique.resultat.filtrage() + chronique.cle() + chronique.resultats.periode
-# Améliorer le fonctionnement en ajoutant une symbolique (3ème version) avec les années/périodes complètes ou incomplètes (parenthèses ou accolades pour les années incomplètes ?)
-#####################
 
 chronique.resultats.periode <- function(
   data = data,
@@ -45,48 +39,51 @@ chronique.resultats.periode <- function(
     {if("Typemesure" %in% colnames(.)) mutate(., paramtypemesure = Typemesure) else .} %>% 
     {if("chres_typemesure" %in% colnames(.)) mutate(., paramtypemesure = chres_typemesure) else .}
 
+  data <-
+    data %>% 
+    chronique.variables.renommage(formatentree = "chres", formatsortie = "param")
+  
   #### Dédoublonnage et tri ####
   dataagrege <- 
     data %>% 
-    {if(nrow(data) != 0) arrange(., paramcoderhj, paramtypemesure, paramannee) else .} %>%
-    distinct(paramcoderhj, paramtypemesure, paramannee) %>% 
-    group_by(paramcoderhj, paramtypemesure)
+    {if(nrow(data) != 0) arrange(., param_coderhj, param_typemesure, param_annee) else .} %>%
+    distinct(param_coderhj, param_typemesure, param_annee) %>% 
+    group_by(param_coderhj, param_typemesure)
 
-  
   #### Calcul ####
   if(typeperiode == "Enveloppe"){
   Vue <-
     dataagrege %>% 
-    summarise(min = min(paramannee),
-              max = max(paramannee)) %>% 
+    summarise(min = min(param_annee),
+              max = max(param_annee)) %>% 
     mutate(PeriodeTotale = paste0(min, " - ", max)) %>% 
     dplyr::select(-min,-max) %>% 
-    left_join(data, by = c("paramcoderhj", "paramtypemesure")) %>% 
+    left_join(data, by = c("param_coderhj", "param_typemesure")) %>% 
     ungroup()
   }
   
   if(typeperiode == "Detail"){
   Vue <-
     dataagrege %>% 
-    mutate(annee = paramannee) %>% 
-    complete(paramannee = min(.$paramannee):max(.$paramannee), nesting(paramcoderhj, paramtypemesure)) %>% 
-    group_by(paramcoderhj, paramtypemesure) %>%
-    mutate(ponctuation = ifelse(lead(annee+1) == annee, ";", "-")) %>%
-    mutate(ponctuation = ifelse(is.na(lead(annee)), ";", "-")) %>%
+    mutate(annee = param_annee) %>% 
+    # complete(param_annee = min(.$param_annee):max(.$param_annee), nesting(param_coderhj, param_typemesure)) %>% # ancienne syntaxe qui buggait, supprimée au 2022-06-08
+    complete(param_annee = min(.$param_annee):max(.$param_annee)) %>% 
+    group_by(param_coderhj, param_typemesure) %>%
+    mutate(ponctuation = ifelse(lead(annee+1) == annee, ",", "-")) %>%
+    mutate(ponctuation = ifelse(is.na(lead(annee)), ",", "-")) %>%
     filter(!is.na(annee)) %>%
-    filter(!(ponctuation == "-" & lag(ponctuation) == "-" & annee != min(paramannee))) %>%
+    filter(!(ponctuation == "-" & lag(ponctuation) == "-" & annee != min(param_annee))) %>%
     mutate(ponctuation = ifelse(annee == max(annee), NA, ponctuation)) %>%
-    mutate(Vue = glue("{paramannee}{ponctuation}", .na = "")) %>%
+    mutate(Vue = glue("{param_annee}{ponctuation}", .na = "")) %>%
     summarise(PeriodeFractionnee = str_c(Vue, collapse = "")) %>% # Agrégation de l'ensemble des valeurs de la colonne en une unique chaîne de caractères
-    left_join(data, by = c("paramcoderhj", "paramtypemesure"))
+    left_join(data, by = c("param_coderhj", "param_typemesure"))
   }
   
   #### Vérification des noms de champs ####
-  # Format standard sans chres_ pour utilisation autonome de l'outil
   Vue <-
     Vue %>% 
-    dplyr::select(-contains("param"))
-  
+    rename_with(~str_replace(., "param_", "chres_"), .cols = everything())
+
   #### Affichage des résultats ####
   return(Vue)
 }